+++ packages/ripple/src/mwc-ripple.ts
@@ -25,6 +25,7 @@ declare global {
   }
 }
 
+/** @soyCompatible */
 @customElement('mwc-ripple')
 export class Ripple extends RippleBase {
   static styles = style;